Modernise pre-commit flow with QOL adjustments
Add flake8-bugbear as a dependency because we were previously
configuring it but not actually using it, so the configuration was
silently ignored.

Add a pydocstyle.ini file which previoulsy wasn't used. The default
pydocstyle config forces you to add docstrings to magic methods (the
so-called double-underscore or data model methods) which have standard,
well-defined meanings, so this wasn't necessary. The parameters of the
__init__ method should be documented under the class's docstring anyway.

Also change the README to restructured text instead of markdown, and
re-wrap the columns to 80 characters. It doesn't make a great deal of
difference when it's rendered on github, but when you're reading the
plain text (which you normally will if you've cloned the repository, and
you're looking at the local copy) then I find that rst makes for easier
reading than md.